% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bgnbd}
\alias{bgnbd}
\title{Estimates the bgnbd model base on “Counting Your Customers” the Easy Way: An Alternative
to the Pareto/NBD Model (2005) by Peter S FaderPeter S FaderBruce G. S. HardieKa Lok Lee}
\usage{
bgnbd(Tx, tx, n, start = rep(1.5, 4), lower = rep(0.01, 4), ...)
}
\arguments{
\item{Tx}{lenght of time since the consumers first purchase.}

\item{tx}{recency of the last purchase (time of the last purchase - time of the first).}

\item{n}{count fo repeat purchases.}

\item{start}{initial values for the bgnbd model.}

\item{lower}{lower argument for optimization defaults to \code{c(0.01, 0.01, 0.01, 0.01)}.}
}
\description{
Estimates the bgnbd model base on “Counting Your Customers” the Easy Way: An Alternative
to the Pareto/NBD Model (2005) by Peter S FaderPeter S FaderBruce G. S. HardieKa Lok Lee
}
\examples{

data(cdnowSample)

## helper function to get the model variables

getMf <- function(set){
   set$week <- as.numeric(set$date - min(set$date)) / 7
   maxWeek <- max(set$week)
   set \%>\%
       group_by(s_id) \%>\% ## for each customer ID
       summarise(
           ## total time being a client
           Tx = maxWeek - min(week),
           ## time between last and first transactions
           tx = max(week) - min(week),
           ## weeks with purchases (ignoring first purchase week)
           n = length(unique(week)) - 1,
           )
}

splitDate <- as.Date("1997-09-30")
calibrateSet <- getMf(filter(cdnowSample, date <= splitDate))
set.seed(345)
model <- bgnbd(Tx = calibrateSet$Tx, tx = calibrateSet$tx,
n = calibrateSet$n, start = c(1, 3, 1, 2.4))
pred <- predict(model, Tx = calibrateSet$Tx, tx = calibrateSet$tx, n = calibrateSet$n, 39)


}
