% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bgnbd_ll}
\alias{bgnbd_ll}
\title{loglike of BGNBD}
\usage{
bgnbd_ll(pars, Tx, tx, n)
}
\arguments{
\item{pars}{(r, alpha, a, b) parameters of the loglike.}

\item{Tx}{lenght of time since the consumers first purchase.}

\item{tx}{recency of the last purchase (time of the last purchase - time of the first)}

\item{n}{count fo repeat purchases}
}
\description{
loglike of BGNBD
}
