% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{cdnowSample}
\alias{cdnowSample}
\title{CDnow dataset sample}
\description{
Dataset pubished by Bruce Hardie see \url{http://www.brucehardie.com/datasets/}.
}
\details{
\enumerate{
  \item \code{m_id} is the client id in the master dataset.
  \item \code{s_id} is the client id.
  \item \code{date} is the date of the purchase
  \item \code{n} is the number of purchased CDs
  \item \code{value} is the dollars payed.
}
}
