% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{palive.bgnbd}
\alias{palive.bgnbd}
\title{Probability of being alive.}
\usage{
\method{palive}{bgnbd}(object, Tx, tx, n)
}
\arguments{
\item{object}{object of class bgnbd}

\item{Tx}{lenght of time since the consumers first purchase.}

\item{tx}{recency of the last purchase (time of the last purchase - time of the first).}

\item{n}{count fo repeat purchases.}

\item{...}{unused}
}
\description{
Palive for bgnbd
}
