% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{palive}
\alias{palive}
\title{Probability of being alive.}
\usage{
palive(object, ...)
}
\arguments{
\item{object}{estimation}

\item{...}{parameters passed to other methods}
}
\description{
generic function to compute the probability of a customer making a purchase in the future.
}
