% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict.bgnbd}
\alias{predict.bgnbd}
\title{Predict expected sales}
\usage{
\method{predict}{bgnbd}(object, Tx, tx, n, t)
}
\arguments{
\item{object}{a object of class \code{bgnbd}.}

\item{Tx}{lenght of time since the consumers first purchase.}

\item{tx}{recency of the last purchase (time of the last purchase - time of the first).}

\item{n}{count fo repeat purchases.}

\item{t}{time period in which the expeted sales take place.}
}
\description{
Expected conditional sales in a \code{t} time period. \eqn{E(Y(t)| X=x, t_x, T, r, \alpha, a, b)}
}
